feat: add API response and pagination test fixtures
* Introduced `apiResponseFixtures` for standardized API response testing.
* Added `paginationTestCases` and `parsePaginationTestCases` for pagination-related tests.
* Updated tests to utilize new fixtures for improved data management and clarity.
